+++ internal/clients/config_status.go
@@ -1,4 +1,4 @@
-package dataplane
+package clients
 
 import (
 	"context"