+++ internal/dataplane/deckgen/generate.go
@@ -121,6 +121,15 @@ func ToDeckContent(
 		return strings.Compare(*content.CACertificates[i].Cert, *content.CACertificates[j].Cert) > 0
 	})
 
+	for _, c := range k8sState.Licenses {
+		content.Licenses = append(content.Licenses,
+			file.FLicense{License: c})
+	}
+
+	sort.SliceStable(content.Licenses, func(i, j int) bool {
+		return strings.Compare(*content.Licenses[i].Payload, *content.Licenses[j].Payload) > 0
+	})
+
 	for _, c := range k8sState.Consumers {
 		consumer := file.FConsumer{Consumer: c.Consumer}