+++ internal/manager/run.go
@@ -18,10 +18,12 @@ import (
 	gatewayv1beta1 "sigs.k8s.io/gateway-api/apis/v1beta1"
 
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/adminapi"
+	"github.com/kong/kubernetes-ingress-controller/v2/internal/clients"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/controllers/gateway"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/dataplane"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/dataplane/sendconfig"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/konnect"
+	"github.com/kong/kubernetes-ingress-controller/v2/internal/license"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/manager/featuregates"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/manager/metadata"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/manager/telemetry"
@@ -108,7 +110,7 @@ func Run(ctx context.Context, c *Config, diagnostic util.ConfigDumpDiagnostic, d
 	setupLog.Info("Initializing Dataplane Client")
 	eventRecorder := mgr.GetEventRecorderFor(KongClientEventRecorderComponentName)
 
-	clientsManager, err := dataplane.NewAdminAPIClientsManager(
+	clientsManager, err := clients.NewAdminAPIClientsManager(
 		ctx,
 		deprecatedLogger,
 		initialKongClients,
@@ -192,6 +194,10 @@ func Run(ctx context.Context, c *Config, diagnostic util.ConfigDumpDiagnostic, d
 	instanceIDProvider := NewInstanceIDProvider()
 
 	if c.Konnect.ConfigSynchronizationEnabled {
+		konnectAPIClient, err := konnect.NewNodeAPIClient(c.Konnect)
+		if err != nil {
+			return fmt.Errorf("failed creating konnect client: %w", err)
+		}
 		// In case of failures when building Konnect related objects, we're not returning errors as Konnect is not
 		// considered critical feature, and it should not break the basic functionality of the controller.
 
@@ -203,6 +209,7 @@ func Run(ctx context.Context, c *Config, diagnostic util.ConfigDumpDiagnostic, d
 		if err := setupKonnectNodeAgentWithMgr(
 			c,
 			mgr,
+			konnectAPIClient,
 			dataplaneClient,
 			clientsManager,
 			setupLog,
@@ -210,6 +217,26 @@ func Run(ctx context.Context, c *Config, diagnostic util.ConfigDumpDiagnostic, d
 		); err != nil {
 			setupLog.Error(err, "Failed to setup Konnect NodeAgent with manager, skipping")
 		}
+	}
+
+	// TODO https://github.com/Kong/kubernetes-ingress-controller/issues/3922
+	// This requires the Konnect client, which currently requires c.Konnect.ConfigSynchronizationEnabled also.
+	// We need to figure out exactly how that config surface works. Initial direction says add a separate toggle, but
+	// we probably want to avoid that long term. If we do have separate toggles, we need an AND condition that sets up
+	// the client and makes it available to all Konnect-related subsystems.
+	if c.Konnect.LicenseSynchronizationEnabled {
+		konnectLicenseAPIClient, err := konnect.NewLicenseAPIClient(c.Konnect)
+		if err != nil {
+			return fmt.Errorf("failed creating konnect client: %w", err)
+		}
+		setupLog.Info("starting license agent")
+		agent := license.NewLicenseAgent(time.Hour*12, "http://example.com", konnectLicenseAPIClient,
+			ctrl.Log.WithName("license-agent"))
+		err = mgr.Add(agent)
+		if err != nil {
+			return fmt.Errorf("could not add license agent to manager: %w", err)
+		}
+		dataplaneClient.EnableLicenseAgent(agent)
 	}
 
 	if c.AnonymousReports {
@@ -245,15 +272,12 @@ func Run(ctx context.Context, c *Config, diagnostic util.ConfigDumpDiagnostic, d
 func setupKonnectNodeAgentWithMgr(
 	c *Config,
 	mgr manager.Manager,
+	konnectNodeAPIClient *konnect.NodeAPIClient,
 	dataplaneClient *dataplane.KongClient,
-	clientsManager *dataplane.AdminAPIClientsManager,
+	clientsManager *clients.AdminAPIClientsManager,
 	logger logr.Logger,
 	instanceIDProvider *InstanceIDProvider,
 ) error {
-	konnectNodeAPIClient, err := konnect.NewNodeAPIClient(c.Konnect)
-	if err != nil {
-		return fmt.Errorf("failed creating konnect client: %w", err)
-	}
 	var hostname string
 	nn, err := util.GetPodNN()
 	if err != nil {
@@ -266,7 +290,7 @@ func setupKonnectNodeAgentWithMgr(
 	version := metadata.Release
 
 	// Set channel to send config status.
-	configStatusNotifier := dataplane.NewChannelConfigNotifier(logger)
+	configStatusNotifier := clients.NewChannelConfigNotifier(logger)
 	dataplaneClient.SetConfigStatusNotifier(configStatusNotifier)
 
 	agent := konnect.NewNodeAgent(
@@ -291,7 +315,7 @@ func setupKonnectNodeAgentWithMgr(
 func setupKonnectAdminAPIClientWithClientsMgr(
 	ctx context.Context,
 	config adminapi.KonnectConfig,
-	clientsManager *dataplane.AdminAPIClientsManager,
+	clientsManager *clients.AdminAPIClientsManager,
 	logger logr.Logger,
 ) {
 	konnectAdminAPIClient, err := adminapi.NewKongClientForKonnectRuntimeGroup(config)