+++ internal/manager/setup.go
@@ -23,6 +23,7 @@ import (
 
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/adminapi"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/admission"
+	"github.com/kong/kubernetes-ingress-controller/v2/internal/clients"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/dataplane"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/manager/scheme"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/util"
@@ -156,7 +157,7 @@ func setupDataplaneSynchronizer(
 func setupAdmissionServer(
 	ctx context.Context,
 	managerConfig *Config,
-	clientsManager *dataplane.AdminAPIClientsManager,
+	clientsManager *clients.AdminAPIClientsManager,
 	managerClient client.Client,
 	deprecatedLogger logrus.FieldLogger,
 ) error {