+++ CHANGELOG.md
@@ -75,6 +75,8 @@ Adding a new version? You'll need three changes:
   configuration. Previously KIC added any UDPRoute that indicated an associated
   Gateway as its parent regardless of Listener configuration or status.
   [#3832](https://github.com/Kong/kubernetes-ingress-controller/pull/3832)
+- Added license agent for Konnect-managed instances.
+  [#3883](https://github.com/Kong/kubernetes-ingress-controller/pull/3883)
 
 ### Fixed